Archived
Moved the blog above the old version of the site in visual hierarchy, and renamed the old site from "Play" to "Old", to fit with what it really is. Might remove it altogether, maybe make it an easter egg or something. That old page needs a lot of work to be presentable.